+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Compilation - Allow to subclass CompiledSchedule in graph compilation (!663).
 - Compilation - Move `MixerCorrections` to `CompilationConfig.hardware_options` (!669)
 - Compilation - Support batched frequencies in schedule resources (!670)
+- Compilation - Move `PowerScaling` (gain/attenuation) to `CompilationConfig.hardware_options` (!673)
 - Operations - The function `convert_to_numerical_pulse` has been added, which can be used to convert `StitchedPulse` to `NumericalPulse` (!665).
   - In addition, this function is called in the pulse diagram plotting functions and the ZI LabOne backend to allow resp. plotting and compiling of `StitchedPulse`.
 - Documentation - Add `UML_class_diagrams` notebook for visualizing class hierarchies (!653)